Move core package functions into compile package.
