+++ core/src/main/scala/quasar/package.scala
@@ -14,102 +14,8 @@
  * limitations under the License.
  */
 
-import slamdata.Predef._
-import quasar.common.{PhaseResult, PhaseResultW}
-import quasar.contrib.pathy._
-import quasar.contrib.scalaz.eitherT._
-import quasar.effect.Failure
-import quasar.fp._
-import quasar.fp.numeric._
-import quasar.frontend.{SemanticErrors, SemanticErrsT}
-import quasar.frontend.logicalplan.{LogicalPlan => LP, Free => _, _}
-import quasar.fs.{CompileM, FileSystemError, FileSystemErrT}
-import quasar.fs.FileSystemError._
-import quasar.fs.PathError._
-import quasar.fs.mount.Mounting
-import quasar.sql._
-import quasar.std.StdLib.set._
-
-import matryoshka._
-import matryoshka.data.Fix
-import matryoshka.implicits._
-import scalaz._, Scalaz._
+import scalaz.EitherT
 
 package object quasar {
-
   type QuasarErrT[M[_], A] = EitherT[M, QuasarError, A]
-
-  private def phase[A: RenderTree](label: String, r: SemanticErrors \/ A):
-      CompileM[A] =
-    EitherT(r.point[PhaseResultW]) flatMap { a =>
-      (a.set(Vector(PhaseResult.tree(label, a)))).liftM[SemanticErrsT]
-    }
-
-  /** Compiles a query into raw LogicalPlan, which has not yet been optimized or
-    * typechecked.
-    */
-  // TODO: Move this into the SQL package, provide a type class for it in core.
-  def precompile[T: Equal: RenderTree]
-    (query: Fix[Sql], vars: Variables, basePath: ADir)
-    (implicit TR: Recursive.Aux[T, LP], TC: Corecursive.Aux[T, LP], S: Show[T])
-      : CompileM[T] = {
-    import SemanticAnalysis._
-    for {
-      ast      <- phase("SQL AST", query.right)
-      substAst <- phase("Variables Substituted", Variables.substVars(ast, vars))
-      absAst   <- phase("Absolutized", substAst.mkPathsAbsolute(basePath).right)
-      sortProj <- phase("Sort Keys Projected", projectSortKeys[Fix[Sql]](absAst).right)
-      annAst   <- phase("Annotated Tree", annotate[Fix[Sql]](sortProj))
-      logical  <- phase("Logical Plan", Compiler.compile[T](annAst) leftMap (_.wrapNel))
-    } yield logical
-  }
-
-  private val optimizer = new Optimizer[Fix[LP]]
-  private val lpr = optimizer.lpr
-
-  /** Optimizes and typechecks a `LogicalPlan` returning the improved plan.
-    */
-  def preparePlan(lp: Fix[LP]): CompileM[Fix[LP]] =
-    for {
-      optimized   <- phase("Optimized", optimizer.optimize(lp).right)
-      typechecked <- lpr.ensureCorrectTypes(optimized) flatMap { a =>
-        (a.set(Vector(PhaseResult.tree("Typechecked", a)))).liftM[SemanticErrsT]
-      }
-      rewritten   <- phase("Rewritten Joins", optimizer.rewriteJoins(typechecked).right)
-    } yield rewritten
-
-  def resolveImports[S[_]](scopedExpr: ScopedExpr[Fix[Sql]], baseDir: ADir)(implicit
-    mount: Mounting.Ops[S],
-    fsFail: Failure.Ops[FileSystemError, S]
-  ): EitherT[Free[S, ?], SemanticError, Fix[Sql]] =
-    EitherT(fsFail.unattemptT(resolveImports_(scopedExpr, baseDir).run))
-
-  def resolveImports_[S[_]](scopedExpr: ScopedExpr[Fix[Sql]], baseDir: ADir)(implicit
-    mount: Mounting.Ops[S]
-  ): EitherT[FileSystemErrT[Free[S, ?], ?], SemanticError, Fix[Sql]] =
-    resolveImportsImpl[EitherT[FileSystemErrT[Free[S, ?], ?], SemanticError, ?], Fix](
-      scopedExpr,
-      baseDir,
-      d => EitherT(EitherT(mount
-             .lookupModuleConfig(d)
-             .bimap(e => SemanticError.genericError(e.shows), _.statements)
-             .run.run ∘ (_ \/> (pathErr(pathNotFound(d))))))
-    ).run >>= (i => EitherT(EitherT.rightT(i.η[Free[S, ?]])))
-
-  /** Returns the `LogicalPlan` for the given SQL^2 query */
-  def queryPlan(
-    expr: Fix[Sql], vars: Variables, basePath: ADir, off: Natural, lim: Option[Positive]):
-      CompileM[Fix[LP]] =
-    precompile[Fix[LP]](expr, vars, basePath)
-      .flatMap(lp => preparePlan(addOffsetLimit(lp, off, lim)))
-
-  def addOffsetLimit[T]
-    (lp: T, off: Natural, lim: Option[Positive])
-    (implicit T: Corecursive.Aux[T, LP])
-      : T = {
-    val skipped = Drop(lp, constant[T](Data.Int(off.value)).embed).embed
-    lim.fold(
-      skipped)(
-      l => Take(skipped, constant[T](Data.Int(l.value)).embed).embed)
-  }
 }
